requirements.txt
gradio
torch
diffusers
transformers
accelerate
fastapi
python-dotenv
Pillow
requests
GPUtil
